Refactor method to delete tool.py
Has only one module to utils/tools method on the library.
